dist/tools/cppcheck: fix all Cppcheck 1.82 errors

+++ pkg/ucglib/src/csrc/ucg_riotos.c
@@ -53,7 +53,7 @@ static void ucg_enable_pins(gpio_t *pins, uint32_t pins_enabled)
     uint8_t i;
 
     for (i = 0; i < 32; i++) {
-        if (pins_enabled & (1 << i)) {
+        if (pins_enabled & ((uint32_t)1 << i)) {
             if (pins[i] != GPIO_UNDEF) {
                 if (i < UCG_PIN_COUNT) {
                     gpio_init(pins[i], GPIO_OUT);
